Added safewrite option using start marker and CRC checks
merged functionality from
https://github.com/AMPWorks/ArduinoLibs/blob/master/EEPromUtils/EEPromUtils.cpp
into the libraries read/write/update block functions. Pre-existing
functions are stil all there and call the new ones using parameters to
switch off new functionality. Also changed readInt in readInt16 and
readLong in to readInt32. This is less confusing as the size of int and
long differ depending on platform and that might lead to mistakes.